+++ pkg/cli/cli_impl.go
@@ -626,10 +626,10 @@ func runImpl(osArgs []string) int {
 
 	switch {
 	case buildOptions != nil:
-		// Read the "NODE_PATH" from the environment. This is part of node's
-		// module resolution algorithm. Documentation for this can be found here:
-		// https://nodejs.org/api/modules.html#modules_loading_from_the_global_folders
 		for _, key := range os.Environ() {
+			// Read the "NODE_PATH" from the environment. This is part of node's
+			// module resolution algorithm. Documentation for this can be found here:
+			// https://nodejs.org/api/modules.html#modules_loading_from_the_global_folders
 			if strings.HasPrefix(key, "NODE_PATH=") {
 				value := key[len("NODE_PATH="):]
 				separator := ":"
@@ -639,6 +639,12 @@ func runImpl(osArgs []string) int {
 				}
 				buildOptions.NodePaths = strings.Split(value, separator)
 				break
+			}
+
+			// Read "NO_COLOR" from the environment. This is a convention that some
+			// software follows. See https://no-color.org/ for more information.
+			if buildOptions.Color == api.ColorIfTerminal && strings.HasPrefix(key, "NO_COLOR=") {
+				buildOptions.Color = api.ColorNever
 			}
 		}